third_party/WebKit/LayoutTests/fast/events/popup-blocking-timers4-expected.txt

Click Here
PASS successfullyParsed is true

TEST COMPLETE
Test calling window.open() in a nested call to setTimeout(). A popup should not be allowed.
PASS newWindow is null
